LayoutTests/imported/w3c/web-platform-tests/cookies/prefix/__host.header.https-expected.txt


PASS __Host: Secure origin: Does not set 'Path=/;'
FAIL __HoSt: Secure origin: Does not set 'Path=/;' assert_equals: expected (undefined) undefined but got (string) "0.35403081743322795"
PASS __Host: Secure origin: Does set 'Secure; Path=/;'
PASS __HoSt: Secure origin: Does set 'Secure; Path=/;'
PASS __Host: Secure origin: Does not set 'Secure; Path=/; Domain=web-platform.test; '
FAIL __HoSt: Secure origin: Does not set 'Secure; Path=/; Domain=web-platform.test; ' assert_equals: expected (undefined) undefined but got (string) "0.3057240585931318"
PASS __Host: Secure origin: Does not set 'Path=/;MaxAge=10'
FAIL __HoSt: Secure origin: Does not set 'Path=/;MaxAge=10' assert_equals: expected (undefined) undefined but got (string) "0.39733369845792965"
PASS __Host: Secure origin: Does set 'Secure; Path=/;MaxAge=10'
PASS __HoSt: Secure origin: Does set 'Secure; Path=/;MaxAge=10'
PASS __Host: Secure origin: Does not set 'Secure; Path=/; Domain=web-platform.test; MaxAge=10'
FAIL __HoSt: Secure origin: Does not set 'Secure; Path=/; Domain=web-platform.test; MaxAge=10' assert_equals: expected (undefined) undefined but got (string) "0.03775975076952098"
PASS __Host: Secure origin: Does not set 'Path=/;HttpOnly'
FAIL __HoSt: Secure origin: Does not set 'Path=/;HttpOnly' assert_equals: expected (undefined) undefined but got (string) "0.43363978431139205"
PASS __Host: Secure origin: Does set 'Secure; Path=/;HttpOnly'
PASS __HoSt: Secure origin: Does set 'Secure; Path=/;HttpOnly'
PASS __Host: Secure origin: Does not set 'Secure; Path=/; Domain=web-platform.test; HttpOnly'
FAIL __HoSt: Secure origin: Does not set 'Secure; Path=/; Domain=web-platform.test; HttpOnly' assert_equals: expected (undefined) undefined but got (string) "0.38190976375285757"
PASS __Host: Secure origin: Does not set 'Secure; Path=/cookies/resources/list.py'
FAIL __HoSt: Secure origin: Does not set 'Secure; Path=/cookies/resources/list.py' assert_equals: expected (undefined) undefined but got (string) "0.9904844508338765"
